+++ BlockV/Core/Network/Stack/API.swift
@@ -204,7 +204,7 @@ extension API {
         ///
         /// - Parameters:
         ///   - message: Content of the message.
-        ///   - userId: Unique identifier of the recipient user.
+        ///   - userID: Unique identifier of the recipient user.
         /// - Returns: The endpoint is generic over a response model. This model is parsed on
         /// success responses (200...299).
         static func sendMessage(_ message: String, toUserId userId: String) -> Endpoint<BaseModel<GeneralModel>> {
@@ -307,7 +307,7 @@ extension API {
         /// Builds the endpoint to search for vAtoms.
         ///
         /// - Parameter payload: Raw request payload.
-        /// - Returns: Endpoint generic over `PackModel`.
+        /// - Returns: Endpoint generic over `UnpackedModel`.
         static func discover(_ payload: [String: Any]) -> Endpoint<BaseModel<UnpackedModel>> {
 
             return Endpoint(method: .post,
@@ -325,7 +325,7 @@ extension API {
         ///   - topRightLat: Top right latitude coordinate.
         ///   - topRightLon: Top right longitude coordinte.
         ///   - filter: The vAtom filter option to apply.
-        /// - Returns: Endpoint generic over `PackModel`.
+        /// - Returns: Endpoint generic over `UnpackedModel`.
         static func geoDiscover(bottomLeftLat: Double,
                                 bottomLeftLon: Double,
                                 topRightLat: Double,